Kali: Edit, Delete and Statistics Functionalities Added And Minor Bugs Fixed

+++ course/views/category.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 from course.models.category import Category
 from course.forms.category import CategoryForm
@@ -11,7 +12,7 @@ def list_categories(request):
     template = "category/list.html"
     categories = Category.objects.all()
     context = {"categories": categories, "list_category_active": "active", "category_show": "show",
-               "category_active": "active"}
+            "category_active": "active"}
     return render(request, template, context)
 
 
@@ -38,11 +39,15 @@ def create_category(request):
 
 @login_required
 def edit_category(request, pk):
-    template = "category/edit.html"
-    category = get_object_or_404(Category, pk=pk)
-    form = CategoryForm(request.GET or None, instance=category)
-    context = {"form": form}
-    return render(request, template, context)
+    if request.user.is_superuser:
+        template = "category/edit.html"
+        category = get_object_or_404(Category, pk=pk)
+        form = CategoryForm(request.GET or None, instance=category)
+        context = {"form": form, "pk": pk}
+        return render(request, template, context)
+    else:
+        raise PermissionDenied(" ")
+
 
 
 @login_required
@@ -55,13 +60,15 @@ def update_category(request, pk):
             messages.success(request, "Category Update Successfully")
             return redirect("list_categories")
         else:
-            messages.error(request, "Category Update Failed!")
-            return redirect("edit_category")
+            raise PermissionDenied("You Are Not An Admin\nPermission Denied")
 
 
 @login_required
 def delete_category(request, pk):
-    category = get_object_or_404(Category, pk=pk)
-    category.delete()
-    messages.success(request, "Rank Deleted Successfully")
-    return redirect("list_categories")
+    if request.user.is_superuser:
+        category = get_object_or_404(Category, pk=pk)
+        category.delete()
+        messages.success(request, "Rank Deleted Successfully")
+        return redirect("list_categories")
+    else:
+        raise PermissionDenied("You Are Not An Admin\nPermission Denied")
