+++ course/views/employee.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 from django_countries.fields import countries
 
@@ -20,10 +21,13 @@ def list_employees(request):
 
 @login_required
 def detail_employees(request, pk):
-    template = "employees/details.html"
-    employee = get_object_or_404(Employee, pk=pk)
-    context = {"employee": employee}
-    return render(request, template, context)
+    if request.user.is_superuser:
+        template = "employees/details.html"
+        employee = get_object_or_404(Employee, pk=pk)
+        context = {"employee": employee}
+        return render(request, template, context)
+    else:
+        raise PermissionDenied("You Are Not An Admin\nPermission Denied")
 
 
 @login_required
@@ -70,32 +74,47 @@ def create_employee(request):
 
 @login_required
 def edit_employee(request, pk):
-    template = "employees/edit.html"
-    employee = get_object_or_404(Employee, pk=pk)
-    employee.trip_set.count()
-    form = EmployeeForm(request.GET or None, instance=employee)
-    context = {"form": form}
-    return render(request, template, context)
+    # import pdb; pdb.set_trace()
+    if request.user.is_superuser:
+        template = "employees/edit.html"
+        employee = get_object_or_404(Employee, pk=pk)
+        # employee.trip_set.count()
+        form = EmployeeForm(request.GET or None, instance=employee)
+        country_edit_code = employee.country.code 
+        selected_category = employee.category
+        categories = Category.objects.all()
+        sex = employee.sex
+        context = {"form": form, "countries": countries, "country_edit_code": country_edit_code, "categories": categories, "sex": sex, "pk": pk}
+        return render(request, template, context)
+    else:
+        raise PermissionDenied("You Are Not An Admin\nPermission Denied")
 
 
 @login_required
 def update_employee(request, pk):
-    if request.method == "POST":
-        employee = get_object_or_404(Employee, pk=pk)
-        form = EmployeeForm(request.POST or None, instance=employee)
-        if form.is_valid():
-            form.save()
-            messages.success(request, "Employee Information Updated Successfully")
-            return redirect("list_employees")
-        else:
-            messages.error(request, "Employee Update Failed")
-            return redirect("edit_employee")
-
+    # import pdb; pdb.set_trace()
+    if request.user.is_superuser:
+        if request.method == "POST":
+            employee = get_object_or_404(Employee, pk=pk)
+            form = EmployeeForm(request.POST or None, instance=employee)
+            # import pdb; pdb.set_trace()
+            if form.is_valid():
+                form.save()
+                messages.success(request, "Employee Information Updated Successfully")
+                return redirect("list_employees")
+            else:
+                messages.error(request, "Employee Update Failed")
+                return redirect("edit_employee")
+    else:
+        raise PermissionDenied("You Are Not An Admin\nPermission Denied")
 
 @login_required
 def delete_employee(request, pk):
-    employee = get_object_or_404(Employee, pk=pk)
-    # form = EmployeeForm(request.GET or None, instance=employee)
-    employee.delete()
-    messages.success(request, "Employee Deleted Successfully!")
-    return redirect("list_employees")
+    if request.user.is_superuser:
+        employee = get_object_or_404(Employee, pk=pk)
+        # form = EmployeeForm(request.GET or None, instance=employee)
+        employee.delete()
+        messages.success(request, "Employee Deleted Successfully!")
+        return redirect("list_employees")
+    else:
+        raise PermissionDenied("You Are Not An Admin\nPermission Denied")
